+++ actix-files/CHANGES.md
@@ -2,6 +2,24 @@
 
 ## Unreleased
 
+- Add `Files::try_compressed()` to support serving pre-compressed static files [#2615]
+- Fix handling of `bytes=0-`
+
+[#2615]: https://github.com/actix/actix-web/pull/2615
+
+## 0.6.10
+
+### Security Notice
+
+We addressed 2 vulnerabilities in this release:
+
+- Do not panic with empty Range header.
+- Avoid serving CWD on invalid `Files::new` inputs.
+
+We encourage updating your `actix-files` version as soon as possible.
+
+### Other changes
+
 - Minimum supported Rust version (MSRV) is now 1.88.
 - `PathBufWrap` & `UriSegmentError` made public. [#3694]